Fix typo in daml-json-types
Actually, it's copy&paste gone wrong. :(

+++ daml-json-types/src/index.ts
@@ -100,7 +100,7 @@ export type Decimal = string;
 /**
  * Companion object of the `Decimal` type.
  */
-export const Decimal: Serializable<Int> = {
+export const Decimal: Serializable<Decimal> = {
   decoder: jtv.string,
 }
